Add second field side painting.
Both field sides can now be painted. To change the field side,
the parameter leftSide in CustomField widget has to be used.
The painting parameter (ellipse radius, gradient and y interception of
section borders) can globally be adapted in fieldSizeParameter.dart.
The check if a click was inside or outside a section works
for both field sides.

Ref Issue #54